[Feat] mapViewModel로 통합 / 차량,탑승유형 MapActivity로 받아오기
