+++ AOS/app/src/main/res/layout/fragment_taxi_information.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
                 name="viewModel"
-                type="org.softeer.robocar.ui.viewmodel.RouteViewModel" />
+                type="org.softeer.robocar.ui.viewmodel.MapViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
